Moved InventoryOwner to item package
