+++ core/src/org/egordorichev/lasttry/item/InventoryOwner.java
@@ -1,4 +1,6 @@
-package org.egordorichev.lasttry.ui;
+package org.egordorichev.lasttry.item;
+
+import org.egordorichev.lasttry.ui.UiInventory;
 
 /**
  * Interface applied to creatures that hold inventories.
