+++ core/src/org/egordorichev/lasttry/ui/UiInventory.java
@@ -11,6 +11,7 @@ import org.egordorichev.lasttry.graphics.Textures;
 import org.egordorichev.lasttry.input.DefaultInputProcessor;
 import org.egordorichev.lasttry.input.InputManager;
 import org.egordorichev.lasttry.input.Keys;
+import org.egordorichev.lasttry.item.InventoryOwner;
 import org.egordorichev.lasttry.item.Item;
 import org.egordorichev.lasttry.item.ItemHolder;
 import org.egordorichev.lasttry.item.modifier.Modifier;
